requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.4.0
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
chromadb==1.0.15
click==8.2.1
coloredlogs==15.0.1
dataclasses-json==0.6.7
dill==0.4.0
distro==1.9.0
docling==2.40.0
docling-core==2.40.0
docling-ibm-models==3.7.0
docling-parse==4.1.0
duckduckgo_search==8.0.5
durationpy==0.10
easyocr==1.7.2
et_xmlfile==2.0.0
fastapi==0.115.14
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.2
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kubernetes==33.1.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-docling==1.0.0
langchain-google-genai==2.1.6
langchain-openai==0.3.27
langchain-pinecone==0.2.8
langchain-tests==0.3.20
langchain-text-splitters==0.3.8
langgraph==0.5.1
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.72
langsmith==0.4.4
latex2mathml==3.78.0
lazy_loader==0.4
lxml==5.4.0
markdown-it-py==3.0.0
marko==2.1.4
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.1.0
mpire==2.10.2
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.18
mypy_extensions==1.1.0
narwhals==1.45.0
networkx==3.5
ninja==1.11.1.4
numpy==2.3.1
oauthlib==3.3.1
onnxruntime==1.22.0
openai==1.93.0
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
orjson==3.10.18
ormsgpack==1.10.0
overrides==7.7.0
packaging==24.2
pandas==2.3.0
pillow==11.3.0
pinecone==7.3.0
pinecone-plugin-assistant==1.7.0
pinecone-plugin-interface==0.0.7
pluggy==1.6.0
posthog==5.4.0
primp==0.15.0
propcache==0.3.2
proto-plus==1.26.1
protobuf>=6.30.0,<7.0.0
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pylatexenc==2.10
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==8.4.1
pytest-asyncio==0.26.0
pytest-benchmark==5.1.0
pytest-codspeed==3.2.0
pytest-recording==0.13.4
pytest-socket==0.7.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
rtree==1.4.0
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.16.0
semchunk==2.2.2
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
streamlit==1.46.1
sympy==1.14.0
syrupy==4.9.1
tabulate==0.9.0
tenacity==9.1.2
tifffile==2025.6.11
tiktoken==0.9.0
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.53.1
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
vcrpy==7.0.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
xlsxwriter==3.2.5
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
